server/requirements.txt
python-decouple
django
djangorestframework
django-rest-auth
django-allauth
django-filter
django-extensions
mysqlclient
Pillow